python-intune-upload/requirements.txt
pyperclip
msgraph-core
azure-identity
webbrowser